Catch and log exceptions that occur in background jobs
When exceptions occur in background jobs that reprocess memes and videos
and such, it kills the background thread that is running them and they
never run again. This commit adds exception handling to those methods
that are called regularly so that they cannot do any harm to the thread
that they are running on.

This has only really been an issue with the YouTube operations so far
but is good practice to add everywhere anyway.