Refresh the meme cache less often
The caches of meme images and video links were updated every hour to
couple of hours. Given how relatively slowly everything that is being
cached actually changes (both the best memes over a long period of time
and the video playlist), this is a huge waste of resources. Every cache
will be updated when the bot starts and then once per day while it is
running, which should be plenty often.